reorg: scalar
