+++ toydl/core/scalar/__init__.py
@@ -1,0 +1,3 @@
+from .scalar import Scalar
+
+__all__ = ["Scalar"]
