--- toydl/util/functions.py
@@ -1,5 +1,0 @@
-def wrap_tuple(x):  # type: ignore
-    """Turn a possible value into a tuple"""
-    if isinstance(x, tuple):
-        return x
-    return (x,)
